requirements.txt
alpaca-trade-api
pandas
numpy
ta
python-dotenv
requests
pytz
python-telegram-bot
alpaca-py>=0.8.0
pandas>=1.5.0
python-dotenv>=1.0.0